requirements.txt
pytest~=8.0.0
allure-pytest==2.13.2
selene~=2.0.0rc4